// src/types/generated/NewNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeKind } from "./NodeKind";

/**
 * Dati per creare un nodo. I campi specifici valgono solo per il loro tipo.
 */
export type NewNode = { kind: NodeKind, name: string, description?: string, icon?: string, colorMain?: string, url?: string, path?: string, };